requirements.txt
python-binance==1.0.15
python-telegram-bot==20.3
pandas==2.0.3
numpy==1.25.2
ta==0.10.2
